backend/requirements.txt
# Python-Abhängigkeiten
django
djangorestframework
psycopg2-binary